refactor(admin): Sichtungstabelle schneiden, Breakpoint vereinheitlichen
Der Seitenkopf von /admin/sichtungen schaltete bei `sm` (640px), die
Inhaltsfläche bei `md` (768px). Zwischen 640 und 767 Pixeln stand damit der
weite Kopf — mit Spalten-Dropdown und Bulk-Kontext — über einer Kartenliste,
die weder Spalten noch Bulk-Auswahl kennt. Betroffen waren Hochkant-Tablets.
`design-system.md` ("Breakpoint-Vertrag") kennt nur `md` und `lg`; `sm` ist
keine Layout-Grenze.

Alles Inhaltliche schaltet jetzt bei `md`: beide Kopf-Varianten, Kartenliste,
Tabelle, die Innenabstände und die Paginierungszeile. Die Datei enthält keine
`sm:`-Klasse mehr.

Damit das nicht erneut auseinanderläuft, steht die Grenze nur noch an einer
Stelle (`layoutSwitch.ts`) und die beiden Layouts sind eigene Komponenten:

- `SichtungenTable.svelte` — Tabelle, Bulk-Leiste, Sortierköpfe, sticky-col
- `SichtungenCards.svelte` — Kartenliste
- `columns.ts` — Default-Sichtbarkeit und Spaltenliste (drei Aufrufstellen)

`+page.svelte` schrumpft dabei von 1960 auf 1307 Zeilen. `actionsColumnWidth`
und `stickyStatusRight` sind komplett in die Tabelle gewandert — beide Seiten
der Messung stehen jetzt in derselben Datei. `groupSuffix="-mobile"` bleibt
unverändert: Beide Layouts stehen weiterhin gleichzeitig im DOM.

Tests: `e2e/admin-table-breakpoint.spec.ts` misst bei 700px und 900px, ob Kopf
und Inhaltsfläche dieselbe Variante zeigen — die Wirkung, nicht die
Klassennamen. Gegenprobe gefahren (Kopf zurück auf `sm` → 700px rot, 900px
grün). `adminPageHeadings.test.ts` hielt die kaputten Klassennamen als Literal
fest und bestätigte den Fehler; er prüft jetzt, dass beide Kopf-Varianten ihre
Sichtbarkeit aus `layoutSwitch.ts` beziehen.
